requirements.txt
# Core web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy[asyncio]>=2.0.25
alembic>=1.13.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Queue and cache
celery>=5.3.0
redis>=5.0.0

# Cloud Storage
aioboto3>=12.3.0
boto3>=1.34.0

# Utils
python-dotenv>=1.0.0
structlog>=24.1.0

# AI
google-generativeai>=0.4.0
assemblyai>=0.33.0

# Dev dependencies
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0
pytest-cov>=4.1.0
auto-editor>=24.1.0

# Render & Notifications
remotion-lambda>=4.0.0
python-telegram-bot>=20.8
